Рефакторинг вспомогательных модулей
